Update hotbit_withdrawal.md
Change MWCMQS to http and Generalize the document to an "Exchange example" which should illustrate a withdraw method instead of a specific Exchange.
Could be deleted completely, but that way some guidance is still there, but all Config Steps that are sepcific to a transaction method are linked in other md's instead of included in here. (to keep it generalized and keep maintenance to a minimum)